✅ Number of Ways to Split Array | Medium | Java | Runtime: 0 ms | Memory: 42.7 MB | 2026-08-16

+++ medium/2270-number-of-ways-to-split-array/solution.java
@@ -1,6 +1,6 @@
 class Solution {
     public int waysToSplitArray(int[] nums) {
-        int[] prefix=new int[nums.length+1];
+        long[] prefix=new long[nums.length+1];
         for(int i=0;i<nums.length;i++){
             prefix[i+1]=prefix[i]+nums[i];
         }
